refactor: move selfie and queries to own files
